Assistant: Support inline editing (#7046)
This contains the minimal changes required to support inline editing in
Positron Assistant. The most important part of the change is supplying a
tool to the model that can perform text edits on any part of the
document; formerly we only had one that could replace the active
selection. We use this new mode when there is no text selected; it has
its own prompt and tool call.

Other notes:

- The system doesn't seem smart enough to avoid using tools that prompt
the user while generating edits. If it tries to use one of these, it
hangs since the model output is not visible during edit generation. Easy
enough to workaround for now but we may need a better approach to this
problem.
- I have removed the R "install packages" tool because it is worse in
every way than just letting the model create an "install.packages()"
call on its own; it was intended to provide a proof of concept as a
state-altering tool with confirmation, but we don't need it any more.

Addresses https://github.com/posit-dev/positron/issues/6959.

## QA Notes

Right now, asking a question in inline chat always dismisses the inline
chat widget. If you asked a question ("explain this") rather than asked
for an edit, it should pop the widget back up to show the answer.