python: status bar path should not be clickable (#7073)
After discussion, it was decided that the python interpreter path in the
status bar should no longer be clickable. We should encourage users
instead to go through the Positron interpreter selector as normal.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- #7001 Python path on status bar no longer pulls up interpreter
selection

#### Bug Fixes

- N/A


### QA Notes

Open Python file.
See interpreter path, click on it. 
Nothing should happen.