Improve reliability of Execute Code tool call (#7082)
This change improves the reliability of Execute Code tool calls in two
ways:

- The code is now run regardless of whether it looks syntactically valid
(complete) or not. If it generates a syntax error, that error will be
sent back to the model to be fixed.
- If the interrupt doesn't work for some reason, the tool call will
still resolve with an error sent to the model, stating that the
computation was interrupted.


![image](https://github.com/user-attachments/assets/69941c7b-93bd-4804-8af2-28bbd8a56db6)

Addresses https://github.com/posit-dev/positron/issues/7080.

---------

Signed-off-by: Jonathan <jonathan@posit.co>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>