Fixes console input queueing (#7105)
Console input queueing support from #6677 was executed 3x, due to the
use of state to track which could hit the case that the runtime changed
state multiple times before the first `Ready` state and the first
re-render. This was changed to use a ref so the flag is not tied to the
render cycle.

---------

Co-authored-by: Chris Mead <chris.mead@posit.co>