Make content emitted from `.Rprofile` display more reliably (#7116)
This change is (another) attempt to address
https://github.com/posit-dev/positron/issues/6344.

The main body of the fix is in the kernel supervisor
(https://github.com/posit-dev/kallichore/pull/30). It was dropping iopub
messages while waiting for a `kernel_info_reply`; these messages are now
saved and sent to the client after the reply has been processed.

Getting this content more reliably introduced a new problem: because
websockets are so fast, it's easy for statements emitted from
`.Rprofile` to reach the client _before_ the kernel info! So you can
wind up in a situation wherein you see your Rprofile's output in the
Console _above_ the startup banner. To help guard against this, I've
added a guard to prevent Jupyter messages from getting streamed until
the kernel is marked ready. There's some risk this could cause other
issues, so we may need to consider a different (read: more complicated)
approach to this problem.

Test tags: `@:console`