Provide attribution to code executions, and an API to listen to them (#7108)
This change adds metadata concerning the provenance of code run in the
Console, so that the code may be given different UX treatment in the
future (e.g. styled to show that it came from an AI assistant, linked
back to the script that ran it, etc.)

It also adds an API that allows extensions to listen to code execution
events. This includes virtually all code run in Positron consoles and
notebooks.

Addresses https://github.com/posit-dev/positron/issues/6689
Addresses https://github.com/posit-dev/positron/issues/2238

### Release Notes


#### New Features

- New API to allow extensions to listen to code execution events (#2238)

#### Bug Fixes

- N/A


### QA Notes

This change doesn't have any visible effects; it just adds an API we can
use in the future.

Test tags: `@:console`