Remove styler R formatter integration (#7199)
Addresses https://github.com/posit-dev/positron/issues/2251
Addresses #19 
Addresses #4435 
Addresses #3916 
Addresses #4009 
Addresses #6017

Prereq for https://github.com/posit-dev/positron/issues/6462

This PR removes the bundled styler integration that came with the
positron-r extension. It effectively reverts #1686.

This immediately improves https://github.com/posit-dev/air/issues/77,
allowing Air to work on Quarto documents in Source Mode in Positron (the
technical details of this one are gross
https://github.com/posit-dev/air/issues/77#issuecomment-2513086286, but
the TLDR is that having 2 formatters installed for the same language
causes havoc in Quarto due to an unfortunate VS Code API decision)

## Next steps

- Our immediate advice after merge should be that people install the Air
extension.
- Once we bootstrap the Air extension
(https://github.com/posit-dev/positron/issues/6462), we should add more
official docs on how to set up Air with #7208, or at least link out to
the Air docs.
- @lionel- I think we should review the fact that `positron-r` sets
`editor.formatOnType: true` for R for indentation rules. Ark is the one
with the LSP handler for
[textDocument/onTypeFormatting](https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_onTypeFormatting),
but it would be pretty reasonable for Air to support "real" on type
formatting (i.e., if the document has no syntax errors, format it), in
which case we'd have a weird kind of conflict with Ark. I recognize that
indentation rules often come into play on "broken" R code though (i.e.
right after a `%>%` with no RHS), so I'm not entirely sure what the
right move is.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- The styler integration for formatting R documents has been removed.
Instead, we encourage you to install the [Air
extension](https://open-vsx.org/extension/posit/air-vscode) for highly
performant R formatting. Positron will bundle the Air extension in a
future release.

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->