sort multiple console dropdown by interpreter type (#7115)
partially addressing #6678

Currently, interpreters are grouped by language and then sorted
alphabetically in the dropdown.. This PR changes it to be grouped by
language > environment type. This is more aligned with the current
`Python: Select Interpreter`/``R: Select Interpreter`` dropdown.

I was playing around with possible ways to make the interpreter display
more logical and reproducible. This adds a level of "skimmability" that
makes interpreters easier to find. It is _not_ using any of the
"[sorting by assumed
usefulness](https://github.com/posit-dev/positron/blob/5ffcabe229a39d339a2a72fdb29cce23d57a5cc3/extensions/positron-python/src/client/interpreter/configuration/environmentTypeComparer.ts#L54)"
that is in the original dropdown. I'm keen to hear other's feedback!

The goal is to:

- within each group, sort by version, then alphabetically

(NOTE: this video is now outdated)


https://github.com/user-attachments/assets/d016af9b-6e30-48fa-9b51-2881373f12d5



### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- 

#### Bug Fixes

- N/A


### QA Notes

Should pass all current tests