Bump ark to v0.1.179 (#7287)
For:

- https://github.com/posit-dev/ark/issues/773 


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Added `ark_positron_variables_has_viewer` to allow packages to
configure the display of the viewer icon in the variables pane.
(https://github.com/posit-dev/positron/issues/7069#issuecomment-2790785343)

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->