Fix hanging runtime on new jupyter notebook project (#7288)
Addresses #5914 

@:new-project-wizard

On new project load sometimes we were trying to run the create
environment command before the python extension had loaded. This then
errored and caused the whole ide to get into a state that looked like it
was perpetually loading the runtime.

This PR adds a function that waits for the python extension to be
initialized before sending the command.

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will
  automatically close the issue. If there are any details about your
  approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- App shouldn't hang initializing runtime on new python projects


### QA Notes

No tests added because this will hopefully fix some existing flakey
tests.