Fix `positron.window.getConsoleForLanguage()` (#7313)
Addresses https://github.com/posit-dev/positron/issues/7169
See
https://github.com/posit-dev/positron/issues/7169#issuecomment-2818886935
for the full story

See also https://github.com/posit-dev/positron/issues/7322 for a follow
up issue we will need to handle, but it is more important to restore the
behavior here first.


https://github.com/user-attachments/assets/4ea98d8b-c0ff-4110-b633-c556ebb9ec1b




### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Manually runnable cli hyperlinks are once again pasted into your
console when you click on them
(https://github.com/posit-dev/positron/issues/7169)


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

@:sessions @:console