Data Explorer: Use complete URIs rather than file paths for DuckDB.
This commit refactors the DuckDB data explorer backend and frontend to
operate on complete URIs rather than filepaths, the first step in
enabling the data explorer to support virtual filesystem providers. It
almost all comes down to swapping out a series of string-typed fields
for vscode.Uri-typed fields.

Since the generated RPC definitions can't currently handle URIs, we
roundtrip through toString() and Uri.parse() at the comm layer.

Note that we still rely on `fs` in the extension, which doesn't work
with non-file URIs, so I added an error message that should surface to
users attempting to open these unsupported URIs. Previously the data
explorer would just be empty when attempting this.

Part of #7351.

Signed-off-by: Aaron Jacobs <aaron.jacobs@posit.co>