change jupyter kernel selector to use Positron's dropdown (#7333)
Updating the jupyter kernel quickpick to use Positron's main interpreter
selector

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Use Positron's main interpreter selector for Jupyter kernels.
Addresses https://github.com/posit-dev/positron/issues/5584

#### Bug Fixes

- N/A


### QA Notes

- Open a Jupyter notebook
- Click on the kernel selector in notebook
- Click `Select Environment...`
- Should see the multiconsole interpreter selector (you should actually
see this selector, even if the multiconsole setting is turned off)