python: resolve paths for fastapi startup (#7383)
So, the native locator will not resolve any path with a tilde; changed
the startup command to use to the untildified path instead.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Fixes a bug where FastAPI apps won't start up with the native Python
locator #5312


### QA Notes

This should work for both js and native locator. Not sure if we need to
turn back on any tests?

```python
from fastapi import FastAPI
app = FastAPI()
@app.get("/")
async def root():
    return {"message": "Hello World"}
```