untildify paths for native locator resolution (#7382)
I've decided to add immediate startup if a user has set
`defaultInterpreterPath` global values, after the feedback in
https://github.com/posit-dev/positron/discussions/6244.

Additionally, the runtime startup sequence was failing with PET, these
paths needed to be untildified. This was causing the active interpreter
(eg, what is started in Jupyter off the bat) to differ from what was the
Positron runtime.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #6280 Current runtime is now selected automatically in Jupyter kernels
outside of a workspace
- #7206 Fixes PET behavior where some environments failed to resolve


### QA Notes