Assistant: Fix "Language model unavailable" issue (#7400)
This change addresses an issue causing "Language model unavailable"
errors when starting chats immediately after registering the first
language model.

The problem is that we always need at least one default language model,
but we only set a default language model when replaying existing
registrations. Consequently the very first model registered doesn't get
marked as a default until the next IDE load.

The fix is to set the `isDefault` flag when the very first model is
registered, so that it can accept requests immediately.

Addresses https://github.com/posit-dev/positron/issues/7391

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->