Support navigating full history when searching (#7376)
This change fixes an issue with history search gestures (e.g.
<kbd>Cmd</kbd> <kbd>Up</kbd>) that caused them to only search the ~10
most recent commands.

The problem is that we use a `HistoryNavigator2` to hold the history
while searching. This object has a default capacity of 10 entries. (You
may be wondering why we used a navigator in the first place, and that's
because an earlier iteration of the search features used a navigator
that had already been constructed for navigating console history).

The fix is to just use the full history array instead of going through a
navigator. Note that this array does not grow unbounded because it is
itself constrained by the console input history setting (default 1000).

Addresses https://github.com/posit-dev/positron/issues/7252.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- The full history can now be searched with Ctrl+Up/Ctrl+R
(https://github.com/posit-dev/positron/issues/7252)


### QA Notes

History is scoped per language, per project, so if you haven't run any
commands in a project yet you won't have any history.

Also note that the search gestures use a history that aggregates all
your consoles (per language), whereas navigation gestures (e.g. up/down)
operate on the history of the current console only.