Fully suppress disabled languages (#7365)
This change increases the strength of the `Disabled` setting for runtime
startup behavior. Previously, it only prevented runtimes from being
_discovered_. Now, it prevents runtimes from being registered _at all_,
even outside the discovery phase.

I noticed that this resulted in a bunch more instances of
https://github.com/posit-dev/positron/issues/6901, so I fixed that while
I was in there. Previously, you were guaranteed to get back a runtime
(or exception) if you called `getPreferredRuntime`. At some call sites,
an error was appropriate, but at other call sites, it isn't. We now
return `undefined` when no suitable runtime is registered, and let the
caller throw their own error if appropriate.

Addresses https://github.com/posit-dev/positron/issues/6268
Addresses https://github.com/posit-dev/positron/issues/6901
Addresses https://github.com/posit-dev/positron/issues/7339

### Release Notes


#### New Features

- N/A

#### Bug Fixes

- Fully disable startup for languages that are disabled in configuration
(https://github.com/posit-dev/positron/issues/7339)


### QA Notes

An easy way to test this is to use a project that has a Python virtual
environment. Python venvs are registered automatically (they do not rely
on discovery) so are an example of a runtime that was previously
registered if Python was disabled, but now should not be.