Fix that allows Enter to be used to select an item in a `DropDownListBox` component (#7405)
## Description

This PR contains a fix that allows Enter to be used to select an item in
a `DropDownListBox` component.

Tests:
@:data-explorer
@:new-project-wizard

### Release Notes

#### New Features

- Partial fix for #4079 that allows Enter to be used to select an item
in a `DropDownListBox` component.

#### Bug Fixes

- N/A

### QA Notes

You can see this in action in two places, in the New Project dialog, and
when adding a row filter in the Data Explorer. This screen recording
shows me adding a row filter using only the keyboard:


https://github.com/user-attachments/assets/900d9421-f9ba-4565-9427-b6fb9fb25b6c