Restarting a busy console now prompts a user to interrupt (#7395)
Now, when a user attempts to restart or delete a busy console, they are
prompted to interrupt first. This allows the action to proceed without
waiting for the slow code to finish or force-quitting.

The prompt allows the user to cancel the action if they don't wish to
discard progress.


Addresses #6881 & #7413