Display pre-renderings of new plots (#7247)
<!-- Thank you for submitting a pull request.
If this is your first pull request you can find information about
contributing here:
  * https://github.com/posit-dev/positron/blob/main/CONTRIBUTING.md

We recommend synchronizing your branch with the latest changes in the
main branch by either pulling or rebasing.
-->

<!--
  Describe briefly what problem this pull request resolves, or what
  new feature it introduces. Include screenshots of any new or altered
  UI. Link to any GitHub issues but avoid "magic" keywords that will
  automatically close the issue. If there are any details about your
  approach that are unintuitive or you want to draw attention to, please
  describe them here.
-->

Addresses https://github.com/posit-dev/positron/issues/5184
Addresses https://github.com/posit-dev/positron/issues/6736
Requires the Ark side PR to be merged first. The Python side still needs
to be implemented.

Plotting currently feels slow (#5184) to users because of the following
back and forth between the frontend and the backend:

- The backend signals there is a plot to display
- The frontend requests a render
- The backend replies with the rendered plot

This setup also prevents plots from being displayed while an execute
request is running since the render request is queued on the Shell
socket and only runs when the execute request has finished. This breaks
some common R patterns where plots are displayed interactively while
taking inputs from users, since this interaction happens in the context
of a single execute request (#6736.

To fix this, the plot widget now supports an optional `pre_render` field
on `comm_open` messages. Backends are allowed to open a plot comm with a
pre-rendering of the plot. This pre-rendering contains the settings used
to generate the plot. If they match what the frontend needs, we're done.
Otherwise we request a new rendering as before but we still display the
pre-render in the meantime. This fixes both issues mentioned above.

Backends can remember the settings of the frontend's last render request
and use that to generate their pre-renderings. With luck these settings
will still be relevant. In the future however, we should send render
settings to interested backends whenever they change (i.e. after the
widget was resized), just like we send console width when the console is
resized.

The comm generator was changed to accomodate a requirement on the Ark
side. The new `RenderPolicy` type needs to be `Copy` so it can be stored
in a `Cell`. This is achieved via a new field `rust.copy` that you can
set on object types. Enum types are always made `Copy` since they will
never contain complex types.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Plots are now displayed faster (#5184).
- Interactive plots (for instance when you call `plot()` on an `lm`
object) are now supported (#6736).

#### Bug Fixes

- N/A


### QA Notes

<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

Since the backends don't get events with up-to-date render settings yet,
the first plot will be displayed right away but it will look a little
bit off since the backend doesn't know which settings to use. It will be
quickly rerendered with the appropriate settings. The second plot will
be displayed with the settings of the last plot which won't be quite
right either since the plot area is resized to make room for the plot
list when a second plot appears in the frontend. However, the third plot
and any plots created after that should be displayed right away with the
correct settings and don't need to be rerendered.

@:plots