Add ability to rename console sessions  (#7384)
Addresses #6825
Addresses #7340 

The `sessionName` field is a user friendly name that can be used in the
UI to identify a session - think of it as a display name. The
`sessionName` field for a session defaults to the `runtimeName` field
from the `runtimeMetaData` object (this is provided by the language
packs). A user couldn't have more than one console session at a time so
the `sessionName` _always_ being the `runtimeName` was sufficient.

Having the `sessionName` field initialized to the `runtimeName` is a
reasonable default that we want to keep. We need to solve the situation
where a user has multiple sessions of the exact same runtime. The first
approach to solve this in #6683 was a _very_ lightweight change. We
introduced a counter to the end of the `sessionName` whenever a user
created more than one session for a given runtime. This change was a
stopgap until we added the ability to rename sessions.

We've gotten enough feedback early on that this functionality needs to
be added sooner rather than later, so this PR does that with the caveat
that the changes aren't as robust as we'd like across the stack
(specifically with getting this change piped up to the API so extensions
know when the session name changes). For example, when restarting a
session, some of the output in the console that comes from an extension
will show the outdated session name. This will mean that there will be a
regression of #7217.

Language packs manage the creation of output channels and they currently
utilize a combination of the `runtimeName`, `sessionName` and/or
`sessionMode` to determine where output should be sent. In the multiple
console session world we are seeing that all output for a given runtime
is sent to the same channel (this is happening even with counters
appended to the `sessionName`. We get two output channels that are both
populated with output data from all sessions for the runtime). This PR
does not resolve the problem of what to do with the output channel name
mismatch that arises when a user renames a session.

A couple things we will want to add in the future:
- An event that extensions can listen to when the session name changes
to keep extensions and positron core in sync
- Configure output channels with different keys to allow session
specific channels

Note: Keyboard actions such as select all/copy/paste are not working in
this PR. These issues will be resolved in #6451 which is addressing the
core issue causing these problems!

### Technical Details

The `sessionName` field has been moved out of the readonly
`RuntimeSessionMetadata` object and into the `LanguageRuntimeDynState`
object. A majority of the files changed in this PR include updating all
the references to `sessionName`.

The `runtimeSession` service includes two additions:
- `updateSessionName` function which handles updating the `sessionName`
- `onDidUpdateSessionName` event that is fired when the `sessionName`
changes. UI components can listen to this event to keep local state up
to date.

There is a new command that allows renaming a session called:
"Interpreter: Rename Console Session" with an id of
`workbench.action.language.runtime.renameSession`

The `consoleTabList` component contains the UI changes that allow a user
to rename a session by right clicking the console tab and selecting the
`Rename` action. This behavior should be equivalent to what the Terminal
pane does.

### Release Notes

#### New Features

- Add ability to rename a console session 

#### Bug Fixes

- N/A

### QA Notes

@:sessions

### Screenshots


https://github.com/user-attachments/assets/55c596f6-d213-4cfc-8063-26986e28adee

**UPDATED CHANGES**


https://github.com/user-attachments/assets/47a5731a-9e25-40bf-8b92-bc33103030b7