Give RStudio keybindings precedence over other contribs (#7456)
When enabling RStudio keyboard shortcuts, a few of them are overridden
by conflicts with other shortcuts in Positron. This is a regression from
the fix for #3993. Shortcuts defined by an extension take precedence
over shortcuts registered in Positron itself; moving the shortcuts from
a builtin extension into Positron core and registering them with the
`WorkbenchContrib` weight (as is typically appropriate for registrations
in `workbench/contrib`) made them vulnerable to override.

The fix is to register these shortcuts as though they were still
contributed by a builtin extension so that they continue to take
precedence over the shortcuts registered by other workbench contribs.

Addresses https://github.com/posit-dev/positron/issues/7434.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- N/A


### QA Notes

Issue #7434 contains a few examples of shortcuts that were affected.