Assistant: Add recent console executions to context (#7454)
This change adds recent console executions to the context provided to
Assistant on every request. This allows you to ask questions about
results or errors in the Console, as well as allowing Assistant to see
the results of code run from the Chat panel (even if you executed it,
not Assistant).

The majority of the new lines here are test code. Adding tests for this
feature was awkward since there were previously no tests for Assistant
core code, and we were missing dependency-free test versions of several
services it depends on, so I've added them (with some help from
generative AI).

Addresses #6964. 

### QA Notes

Only the most recent console executions are sent to Assistant to avoid
making the context window too long, so if you run code that generates a
lot of output, Assistant won't be able to see more than a few entries
back.

Test tags: `@:assistant`